SMS-255

+++ usermanagement-service/src/main/java/com/sms/usermanagementservice/users/control/UsersService.java
@@ -7,6 +7,7 @@ import com.sms.api.usermanagement.CustomAttributesDTO;
 import com.sms.api.usermanagement.UserDTO;
 import com.sms.api.usermanagement.UsersFiltersDTO;
 import com.sms.usermanagementservice.clients.GradesClient;
+import com.sms.usermanagementservice.clients.HomeworksClient;
 import com.sms.usermanagementservice.users.entity.CustomFilterParams;
 import com.sms.usermanagementservice.users.entity.KeyCloakFilterParams;
 import org.keycloak.representations.idm.UserRepresentation;
@@ -33,6 +34,10 @@ public class UsersService {
 
     @Autowired
     GradesClient gradesClient;
+
+    @Autowired
+    HomeworksClient homeworksClient;
+
 
     public List<UserDTO> filterUserByParameters(UsersFiltersDTO filterParamsDTO) {
         CustomFilterParams customFilterParams = UserMapper.mapCustomFilterParams(filterParamsDTO);
@@ -71,7 +76,7 @@ public class UsersService {
         UserRepresentation userRepresentation = keycloakClient.getUser(userId)
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
 
-        if (gradesClient.deleteGrades(userRepresentation.getId())) {
+        if (gradesClient.deleteGrades(userRepresentation.getId()) && homeworksClient.deleteAnswers(userRepresentation.getId())) {
             Boolean isDeleted = deleteRelatedUser(userRepresentation);
             if (!(keycloakClient.deleteUser(userId) && isDeleted)) {
                 throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR);
